langchain/RAG/requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
beautifulsoup4